Improving handling of failed video file analysis.

+++ net.jancerveny.sofaking.BusinessLogic/Models/EncoderConfiguration.cs
@@ -9,8 +9,8 @@ namespace net.jancerveny.sofaking.BusinessLogic.Models
 		public string FFMPEGBinary { get; set; }
 		public string OutputVideoCodec { get; set; }
 		public string OutputAudioCodec { get; set; }
-		public int OutputVideoBitrateMbits { get; set; }
-		public int OutputAudioBitrateMbits { get; set; }
+		public double OutputVideoBitrateMbits { get; set; }
+		public double OutputAudioBitrateMbits { get; set; }
 		public string TempFolder { get; set; }
 	}
 }
